Day 10

+++ 2022/day08.py
@@ -21,7 +21,7 @@ if __name__ == '__main__':
     X, Y = stand.shape
     
     visible = np.zeros((X, Y), dtype=bool) # boolean array for tree visibility
-    scenic = np.zeros((X, Y))
+    scenic = np.zeros((X, Y), dtype=int)
     
     for i in range(X):
         for j in range(Y):
@@ -31,8 +31,8 @@ if __name__ == '__main__':
             arrs.append(np.array(stand[i, j+1:])) # look down
             arrs.append(np.flip(np.array(stand[:i, j]))) # look left
             arrs.append(np.array(stand[i+1:, j])) # look right
-            visible[i, j] = np.sum([np.all( val > x) for x in arrs])
+            visible[i, j] = np.sum([np.all(val > x) for x in arrs])
             scenic[i, j] = np.prod([get_viewing_distance(val, x) for x in arrs])
     
-    print('part 1:', int(np.sum(visible)))
-    print('part 2:', int(np.max(scenic)))
+    print('part 1:', np.sum(visible))
+    print('part 2:', np.max(scenic))
